+++ crates/radicle-cli/src/commands/id.rs
@@ -19,6 +19,7 @@ use crate::git::unified_diff::Encode as _;
 use crate::git::Rev;
 use crate::terminal as term;
 use crate::terminal::args::Error;
+use crate::terminal::format::Author;
 use crate::terminal::patch::Message;
 
 pub use args::Args;
@@ -193,7 +194,14 @@ pub fn run(args: Args, ctx: impl term::Context) -> anyhow::Result<()> {
                 return Ok(());
             }
             let signer = term::signer(&profile)?;
-            let revision = update(title, description, proposal, &mut identity, &signer)?;
+            let revision = update(
+                title,
+                description,
+                proposal,
+                &mut identity,
+                &signer,
+                &profile,
+            )?;
 
             if revision.is_accepted() && revision.parent == Some(current.id) {
                 // Update the canonical head to point to the latest accepted revision.
@@ -420,13 +428,16 @@ fn update<R, G>(
     doc: Doc,
     current: &mut IdentityMut<R>,
     signer: &Device<G>,
+    profile: &Profile,
 ) -> anyhow::Result<Revision>
 where
     R: WriteRepository + cob::Store<Namespace = NodeId>,
     G: crypto::signature::Signer<crypto::Signature>,
 {
     if let Some((title, description)) = edit_title_description(title, description)? {
-        let id = current.update(title, description, &doc, signer)?;
+        let id = current
+            .update(title, description, &doc, signer)
+            .map_err(|e| on_identity_err(e, profile))?;
         let revision = current
             .revision(&id)
             .ok_or(anyhow!("update failed: revision {id} is missing"))?;
@@ -434,6 +445,46 @@ where
         Ok(revision.clone())
     } else {
         Err(anyhow!("you must provide a revision title and description"))
+    }
+}
+
+fn on_identity_err(e: identity::Error, profile: &Profile) -> anyhow::Error {
+    let e = anyhow::Error::from(e);
+
+    e.chain()
+        .find_map(|c| c.downcast_ref::<identity::ApplyError>())
+        .map(|e| on_apply_err(e, profile))
+        .unwrap_or(e)
+}
+
+fn on_apply_err(e: &identity::ApplyError, profile: &Profile) -> anyhow::Error {
+    match e {
+        e @ identity::ApplyError::NonDelegateUnauthorized { author, .. } => {
+            let nid = NodeId::from(*author);
+            let labels = Author::new(&nid, profile, false).labels();
+
+            Error::with_hint(
+                anyhow!(e.to_string()),
+                format!(
+                    "{} {} is attempting to modify the identity document but is not a delegate!",
+                    labels.0, labels.1
+                ),
+            )
+            .into()
+        }
+        e @ radicle::cob::identity::ApplyError::Missing(_)
+        | e @ radicle::cob::identity::ApplyError::Init(_)
+        | e @ radicle::cob::identity::ApplyError::InvalidSignature(..)
+        | e @ radicle::cob::identity::ApplyError::NotAuthorized
+        | e @ radicle::cob::identity::ApplyError::MissingParent
+        | e @ radicle::cob::identity::ApplyError::DuplicateVerdict
+        | e @ radicle::cob::identity::ApplyError::UnexpectedState
+        | e @ radicle::cob::identity::ApplyError::Redacted
+        | e @ radicle::cob::identity::ApplyError::DocUnchanged
+        | e @ radicle::cob::identity::ApplyError::Git(_)
+        | e @ radicle::cob::identity::ApplyError::Doc(_) => {
+            anyhow!(e.to_string())
+        }
     }
 }